gcs-synapse-sync/requirements.txt
synapseclient>=2.1,<3
google-cloud-secret-manager>=2.4.0
google-cloud-storage>=1.37.0
google-api-python-client>=2.0.2
